+++ JSUtils.swift
@@ -1,0 +1,73 @@
+//
+//  JSUtils.swift
+//  PMKJSA+Tests
+//
+//  Created by Lois Di Qual on 3/2/18.
+//
+
+import Foundation
+import JavaScriptCore
+
+enum JSUtils {
+    
+    class JSError: Error {
+        let reason: JSValue
+        init(reason: JSValue) {
+            self.reason = reason
+        }
+    }
+    
+    static let sharedContext: JSContext = {
+        guard let context = JSContext() else {
+            fatalError("Couldn't create JS context")
+        }
+        return context
+    }()
+    
+    static var undefined: JSValue {
+        guard let undefined = JSValue(undefinedIn: JSUtils.sharedContext) else {
+            fatalError("Couldn't create `undefined` value")
+        }
+        return undefined
+    }
+    
+    // @warning: relies on lodash to be present
+    static func isFunction(value: JSValue) -> Bool {
+        guard let context = value.context else {
+            return false
+        }
+        guard let lodash = context.objectForKeyedSubscript("_") else {
+            fatalError("Couldn't get lodash in JS context")
+        }
+        guard let result = lodash.invokeMethod("isFunction", withArguments: [value]) else {
+            fatalError("Couldn't invoke _.isFunction")
+        }
+        return result.toBool()
+    }
+    
+    // Calls a JS function using `Function.prototype.call` and throws any potential exception wrapped in a JSError
+    static func call(function: JSValue, arguments: [JSValue]) throws -> JSValue? {
+        
+        let context = JSUtils.sharedContext
+        
+        // Create a new exception handler that will store a potential exception
+        // thrown in the handler. Save the value of the old handler.
+        var caughtException: JSValue?
+        let savedExceptionHandler = context.exceptionHandler
+        context.exceptionHandler = { context, exception in
+            caughtException = exception
+        }
+        
+        // Call the handler
+        let returnValue = function.invokeMethod("call", withArguments: arguments)
+        context.exceptionHandler = savedExceptionHandler
+        
+        // If an exception was caught, throw it
+        if let exception = caughtException {
+            throw JSError(reason: exception)
+        }
+        
+        return returnValue
+    }
+    
+}
